# packages/api-client/openapi.json
{
...
  "info": {
    "contact": {},
    "description": "REST API for the Hà Tĩnh immersive tourism platform foundation.",
    "title": "Hà Tĩnh Immersive API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/admin/destinations": {
      "post": {
        "operationId": "createDestination",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "categoryId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  },
                  "coverMediaId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  },
                  "defaultSceneId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  },
                  "geoPoint": {
                    "nullable": true,
                    "properties": {
                      "latitude": {
                        "example": 18.3421,
                        "type": "number"
                      },
                      "longitude": {
                        "example": 105.9032,
                        "type": "number"
                      }
                    },
                    "type": "object"
                  },
                  "slug": {
                    "example": "son-trang-co-dam",
                    "type": "string"
                  },
                  "translations": {
                    "items": {
                      "properties": {
                        "description": {
                          "type": "string"
                        },
                        "locale": {
                          "example": "vi",
                          "type": "string"
                        },
                        "name": {
                          "example": "Sơn Trang Cổ Đạm",
                          "type": "string"
                        },
                        "summary": {
                          "type": "string"
                        }
                      },
                      "required": ["locale", "name", "summary"],
                      "type": "object"
                    },
                    "type": "array"
                  }
                },
                "required": ["slug", "translations"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "categoryId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "coverMediaId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "defaultSceneId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "geoPoint": {
                      "nullable": true,
                      "properties": {
                        "latitude": {
                          "example": 18.3421,
                          "type": "number"
                        },
                        "longitude": {
                          "example": 105.9032,
                          "type": "number"
                        }
                      },
                      "required": ["latitude", "longitude"],
                      "type": "object"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "status": {
                      "enum": ["draft", "published", "archived"],
                      "type": "string"
                    },
                    "translations": {
                      "items": {
                        "properties": {
                          "description": {
                            "type": "string"
                          },
                          "locale": {
                            "example": "vi",
                            "type": "string"
                          },
                          "name": {
                            "example": "Sơn Trang Cổ Đạm",
                            "type": "string"
                          },
                          "summary": {
                            "type": "string"
                          }
                        },
                        "required": ["locale", "name", "summary", "description"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "slug",
                    "status",
                    "categoryId",
                    "geoPoint",
                    "defaultSceneId",
                    "coverMediaId",
                    "translations",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Created destination draft."
          }
        },
        "summary": "",
        "tags": ["admin-destinations"]
      }
    },
    "/api/v1/admin/destinations/{id}": {
      "patch": {
        "operationId": "updateDestination",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "categoryId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  },
                  "coverMediaId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  },
                  "defaultSceneId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  },
                  "geoPoint": {
                    "nullable": true,
                    "properties": {
                      "latitude": {
                        "example": 18.3421,
                        "type": "number"
                      },
                      "longitude": {
                        "example": 105.9032,
                        "type": "number"
                      }
                    },
                    "type": "object"
                  },
                  "slug": {
                    "example": "son-trang-co-dam",
                    "type": "string"
                  },
                  "translations": {
                    "items": {
                      "properties": {
                        "description": {
                          "type": "string"
                        },
                        "locale": {
                          "example": "vi",
                          "type": "string"
                        },
                        "name": {
                          "example": "Sơn Trang Cổ Đạm",
                          "type": "string"
                        },
                        "summary": {
                          "type": "string"
                        }
                      },
                      "required": ["locale", "name", "summary"],
                      "type": "object"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "categoryId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "coverMediaId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "defaultSceneId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "geoPoint": {
                      "nullable": true,
                      "properties": {
                        "latitude": {
                          "example": 18.3421,
                          "type": "number"
                        },
                        "longitude": {
                          "example": 105.9032,
                          "type": "number"
                        }
                      },
                      "required": ["latitude", "longitude"],
                      "type": "object"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "status": {
                      "enum": ["draft", "published", "archived"],
                      "type": "string"
                    },
                    "translations": {
                      "items": {
                        "properties": {
                          "description": {
                            "type": "string"
                          },
                          "locale": {
                            "example": "vi",
                            "type": "string"
                          },
                          "name": {
                            "example": "Sơn Trang Cổ Đạm",
                            "type": "string"
                          },
                          "summary": {
                            "type": "string"
                          }
                        },
                        "required": ["locale", "name", "summary", "description"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "slug",
                    "status",
                    "categoryId",
                    "geoPoint",
                    "defaultSceneId",
                    "coverMediaId",
                    "translations",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Updated destination draft."
          }
        },
        "summary": "",
        "tags": ["admin-destinations"]
      }
    },
    "/api/v1/admin/destinations/{id}/publish": {
      "post": {
        "operationId": "publishDestination",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "categoryId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "coverMediaId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "defaultSceneId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "geoPoint": {
                      "nullable": true,
                      "properties": {
                        "latitude": {
                          "example": 18.3421,
                          "type": "number"
                        },
                        "longitude": {
                          "example": 105.9032,
                          "type": "number"
                        }
                      },
                      "required": ["latitude", "longitude"],
                      "type": "object"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "status": {
                      "enum": ["draft", "published", "archived"],
                      "type": "string"
                    },
                    "translations": {
                      "items": {
                        "properties": {
                          "description": {
                            "type": "string"
                          },
                          "locale": {
                            "example": "vi",
                            "type": "string"
                          },
                          "name": {
                            "example": "Sơn Trang Cổ Đạm",
                            "type": "string"
                          },
                          "summary": {
                            "type": "string"
                          }
                        },
                        "required": ["locale", "name", "summary", "description"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "slug",
                    "status",
                    "categoryId",
                    "geoPoint",
                    "defaultSceneId",
                    "coverMediaId",
                    "translations",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Published destination."
          }
        },
        "summary": "",
        "tags": ["admin-destinations"]
      }
    },
...
  },
  "servers": [],
...
}